Move NodeOptions and OccupancyGrid building.
Constructing NodeOptions and nav_msgs::OccupancyGrids
is moved out of the ROS node implementation file.
